% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.jq_date_oldest_id}
\alias{.jq_date_oldest_id}
\title{jq filter to extract create date of oldest Tweet in v2 search endpoint JSON}
\usage{
.jq_date_oldest_id(tweetId)
}
\arguments{
\item{tweetId}{a character string specifying the ID of the oldest Tweet
contained in a v2 search endpoint JSON reponse; typcially obtained via
\code{.oldest_id} and \code{.search_metadata}}
}
\description{
jq filter to extract create date of oldest Tweet in v2 search endpoint JSON
}
\keyword{internal}
