% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.jq_metadata_v2_filter}
\alias{.jq_metadata_v2_filter}
\title{jq filter to extract response metadata for pagination}
\usage{
.jq_metadata_v2_filter()
}
\description{
jq filter to extract response metadata for pagination
}
\keyword{internal}
