% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.newest_id}
\alias{.newest_id}
\title{ID of the most recent Tweet in a given result set}
\usage{
.newest_id(searchMetadata)
}
\arguments{
\item{searchMetadata}{the search metadata dataframe returned by
\code{.search_metadata()}}
}
\description{
ID of the most recent Tweet in a given result set
}
\keyword{internal}
