% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.next_token}
\alias{.next_token}
\title{Search pagination token a given result set}
\usage{
.next_token(searchMetadata)
}
\arguments{
\item{searchMetadata}{the search metadata dataframe returned by
\code{.search_metadata()}}
}
\description{
The \code{next_token} parameter is required to traverse a result set with
consecutive API calls.
}
\keyword{internal}
