% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.oldest_id}
\alias{.oldest_id}
\title{ID of the oldest Tweet in a given result set}
\usage{
.oldest_id(searchMetadata)
}
\arguments{
\item{searchMetadata}{the search metadata dataframe returned by
\code{.search_metadata()}}
}
\description{
ID of the oldest Tweet in a given result set
}
\keyword{internal}
