% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.oldest_tweet_date}
\alias{.oldest_tweet_date}
\title{Create date of the oldest Tweet in a v2 search as POSIXct date-time object}
\usage{
.oldest_tweet_date(jsonSource, tweetId)
}
\arguments{
\item{jsonSource}{a JSON string as returned by the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint}

\item{tweetId}{a character string specifying the ID of the oldest Tweet
contained in a v2 search endpoint JSON reponse; typcially obtained via
\code{.oldest_id} and \code{.search_metadata}}
}
\description{
Create date of the oldest Tweet in a v2 search as POSIXct date-time object
}
\keyword{internal}
