% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.result_count}
\alias{.result_count}
\title{Number of Tweets in a given result set}
\usage{
.result_count(searchMetadata)
}
\arguments{
\item{searchMetadata}{the search metadata dataframe returned by
\code{.search_metadata()}}
}
\description{
Number of Tweets in a given result set
}
\keyword{internal}
