% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.search_metadata}
\alias{.search_metadata}
\title{Search metadata as a dataframe}
\usage{
.search_metadata(jsonSource)
}
\arguments{
\item{jsonSource}{a JSON string as returned by the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint}
}
\description{
Search metadata as a dataframe
}
\keyword{internal}
