% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-tweets.R
\name{.url_v2_search_tweets_all}
\alias{.url_v2_search_tweets_all}
\title{Create parameterised URL for v2 API full archive Tweet search}
\usage{
.url_v2_search_tweets_all(queryString, maxResult = 500, nextToken = NA,
  fromDate = NULL, toDate = NULL)
}
\description{
Create parameterised URL for v2 API full archive Tweet search
}
\keyword{internal}
