% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-tweets.R
\name{get_v2_tweets_search_all}
\alias{get_v2_tweets_search_all}
\title{Get API response for a single full archive Tweet search query}
\usage{
get_v2_tweets_search_all(queryString, fromDate = NULL, toDate = NULL,
  nextToken = NA, maxResult = 500, twitterBearerToken)
}
\value{
the complete Twitter API response
}
\description{
Execute a Tweet search against the the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint and return the HTTP response object.
}
\details{
Same parameters as \code{search_tweets()}.
}
\keyword{internal}
