% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{oauth_twitter_token}
\alias{oauth_twitter_token}
\title{Authenticate Twitter API access and get bearer token}
\usage{
oauth_twitter_token(consumerKey, consumerSecret)
}
\arguments{
\item{consumerKey}{a character string specifying the consumer key for the
registered authenticating Twitter application in the Academic Research
product track}

\item{consumerSecret}{a character string specifying the consumer secret for
the registered authenticating Twitter application in the Academic Research
product track}
}
\value{
a character string representing a valid bearer token
}
\description{
\code{oauth_twitter_token} authenticates the provided credentials against the
Twitter
\href{https://developer.twitter.com/en/docs/authentication/api-reference/token}{\code{oauth2/token}}
endpoint and returns the authentication (bearer) token required for calls to
other Twitter API endpoints (consult the Twitter API documentation on
\href{https://developer.twitter.com/en/docs/authentication/oauth-2-0/application-only}{application-only
authentication} for details).
}
\details{
This authentication method requires a \strong{consumer key} and
\strong{consumer secret} for a registered Twitter application (see
\href{https://developer.twitter.com/en/docs/apps/overview}{here} for
details).
}
