% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-tweets.R
\name{tweets_search_and_store}
\alias{tweets_search_and_store}
\title{Retrieve and store all Tweets for a given full archive search query}
\usage{
tweets_search_and_store(queryString, fromDate = NULL, toDate = NULL,
  batchBaseLabel = NULL, maxBatchSize = 10000,
  twitterBearerToken = NULL)
}
\arguments{
\item{queryString}{a character string specifying a value for the Tweet search
query parameter (e.g. \emph{"sustainability (climate change)"},
\emph{"from:stefandaume"} etc). See
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{here}
for details. Maximum of 1024 characters.}

\item{fromDate}{a character string of format ("YYYY-MM-DD") specifying the
date for the oldest Tweets to be included in the search results
(interpreted as inclusive); corresponds to the \code{start_time} parameter
of the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint. Must NOT be a date before "2006-03-21". If no value is supplied
(default), it is interpreted as the 30th day before \code{toDate}.}

\item{toDate}{a character string of format ("YYYY-MM-DD") specifying the date
for the most recent Tweets to be included in the search results
(interpreted as inclusive); corresponds to the \code{end_time} parameter of
the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint. If no value is supplied (default), it is interpreted as the
current date.}

\item{batchBaseLabel}{a character string used to name the stored Tweet search
batch files.}

\item{maxBatchSize}{an integer specifying the approximate number of Tweets in
each stored batch of Tweet search responses. This number serves as a
threshold, once more Tweets than \code{maxBatchSize} are accumulated, they
are stored as a batch.}

\item{twitterBearerToken}{a character string specifying a valid bearer token
for the
\href{https://developer.twitter.com/en/solutions/academic-research}{Twitter
Academic Research product track} (see \code{oauth_twitter_token()}).}
}
\description{
\code{tweets_search_and_store} allows to execute long-running Tweet searches
against the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint in the Twitter Academic Research product track
\href{https://developer.twitter.com/en/solutions/academic-research}{Twitter
Academic Research product track} and store the results as batches of the
original JSON responses.
}
\details{
Based on a given Tweet search query and (optional) date range this function
iteratively retrieves all matching Tweets and stores the results in batches
as appropriately labelled JSON files; the complete set of results is split in
batches containing the (approximate) number of Tweets specified by
\code{maxBatchSize} Tweets.

The calls to the search API endpoint are timed such the API call limit of at
most 300 calls per 15 minute window is observed; this corresponds to a
maximum of 150.000 Tweets that can be retrieved every 15 minutes (each
individual API call returns at most 500 Tweets).
}
