% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twittrcademic.R
\docType{package}
\name{twittrcademic}
\alias{twittrcademic}
\alias{twittrcademic-package}
\title{twittrcademic: Collecting Tweet Data Via The Twitter Academic Research
Product Track}
\description{
This package provides a collection of utility functions supporting the
collection of Tweet data via Twitter v2 API endpoints in the Twitter Academic
Research product track
\href{https://developer.twitter.com/en/solutions/academic-research}{Twitter
Academic Research product track}.
}
