% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-tweets.R
\name{v2_tweets_search_all}
\alias{v2_tweets_search_all}
\title{Get JSON response for a single full archive Tweet search query}
\usage{
v2_tweets_search_all(queryString, fromDate = NULL, toDate = NULL,
  nextToken = NA, maxResult = 500, twitterBearerToken)
}
\arguments{
\item{fromDate}{a character string of format ("YYYY-MM-DD") specifying the
date for the oldest Tweets to be included in the search result (interpreted
as inclusive); corresponds to the \code{start_time} parameter of the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint. Must NOT be a date before "2006-03-21". If no value is supplied
(default), it is interpreted as the 30th day before \code{toDate}.}

\item{toDate}{a character string of format ("YYYY-MM-DD") specifying the date
for the most recent Tweets to be included in the search result (interpreted
as inclusive); corresponds to the \code{end_time} parameter of the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint. If no value is supplied (default), it is interpreted as the
current date.}

\item{nextToken}{a character string specifying a token used to iterate over
search results larger than the maximum of 500. If a Tweet search has more
results, the metadata section in the JSON response will provide a
\strong{next_token} value, which can be used to paginate over the next set
of results (see
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/integrate/paginate}{here}
for details.) \code{tweets_search_and_store()} uses this mechanism to
retrieve complete result sets.}

\item{maxResult}{an integer specifying the maximum number of Tweets returned
in a single search API call (minimum is 10, maximum is 500)}

\item{twitterBearerToken}{a character string specifying a valid bearer token
for the
\href{https://developer.twitter.com/en/solutions/academic-research}{Twitter
Academic Research product track} (see \code{oauth_twitter_token()}).}
}
\value{
the Twitter API JSON response as a JSON character string
}
\description{
\code{v2_tweets_search_all} allows to execute a single Tweet search API call
against the
\href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{/2/tweets/search/all}
endpoint in the Twitter Academic Research product track.
}
